The player now spawns in the next room properly.
The player spawns at the spawn point accessed through the name of the
level the player just left.  This means that currently there can only be
one entrance to each level for one room.  I will change this later so
that the SpawnKey is obtained from the LevelExit rectangle that the
player used.